+++ __test__/noteoverview.test.ts
@@ -92,6 +92,10 @@ describe("Get image nr X from body", function () {
         ![ad762c6793d46b521cea4b2bf3f01b5e.png](:/a7f9ed618c6d427395d1ef1db2ee2000)
         text
         ![](:/766bf08661e51d3897e6314b56f4d113)  
+        text
+        <img src=":/a1fd1b6fd6be4ab58f99e01beb704b18" alt="1aa911358498d84e725fba441e88c05a.png" width="392" height="246">
+        text
+        <img  alt='8f99e01beb704b18a1fd1b6fd6be4ab5.png' width='392' src=':/8f99e01beb704b18a1fd1b6fd6be4ab5' height='246'>
         `;
 
     imgStr = await noteoverview.getImageNr(body, 1, imageSettings);
@@ -104,8 +108,18 @@ describe("Get image nr X from body", function () {
       `<img src=':/766bf08661e51d3897e6314b56f4d113' width='200' height='200'>`
     );
 
-    imgStr = await noteoverview.getImageNr(body, 4, imageSettings);
+    imgStr = await noteoverview.getImageNr(body, 6, imageSettings);
     expect(imgStr).toBe(``);
+
+    imgStr = await noteoverview.getImageNr(body, 4, imageSettings);
+    expect(imgStr).toBe(
+      `<img src=':/a1fd1b6fd6be4ab58f99e01beb704b18' width='200' height='200'>`
+    );
+
+    imgStr = await noteoverview.getImageNr(body, 5, imageSettings);
+    expect(imgStr).toBe(
+      `<img src=':/8f99e01beb704b18a1fd1b6fd6be4ab5' width='200' height='200'>`
+    );
   });
 
   it(`with settings`, async () => {